bug 773943 - fix null pointer exception from accessing getExternalFilesDir r=snorp
UltraBlame original commit: 8f93bcc5bc5662ee36df98053209ad094316b364

+++ mobile/android/base/GeckoApp.java
@@ -2377,16 +2377,17 @@ abstract public class GeckoApp
     
     public static File getTempDirectory() {
         File dir = mAppContext.getExternalFilesDir("temp");
-        dir.mkdirs();
         return dir;
     }
 
     
     public static void deleteTempFiles() {
-        File[] files  = getTempDirectory().listFiles();
+        File dir = getTempDirectory();
+        if (dir == null)
+            return;
+        File[] files = dir.listFiles();
         if (files == null)
             return;
-
         for (File file : files) {
             file.delete();
         }
